ExternalBallistics.jl Documentation

Trajectory computation.

Introduction

julia> using ExternalBallistics

Package Features

MPMM direct and indirect

Functions Documentation

Objects generation

ExternalBallistics.createTargetRect — Function

creates a rectangular target by specifying the length a, the height b and the position (range)

ExternalBallistics.createTargetCirc — Function

creates a circular target by specifying the radius ρ and the position (range)

ExternalBallistics.createTarget — Function

Creates a target. Optional options are the size and the position

ExternalBallistics.createGun — Function

Creates a Gun by specifying the muzzle velocity its latitude, the elevation angle, azimuth angle and the twist. Optional options are the gun length and the height of the gun

Trajectory computation

ExternalBallistics.trajectoryMPMM — Function

computes the impact point, the velocity at impact and the time of flight by providing a projectile, a target, a gun and aerodynamic coefficient data.
optional options are the computational time, the constant R, the constant Ω, the wind velocity and the atmosphere.
The trajectory is computed using MPMM

ExternalBallistics.QEfinderMPMM! — Function

computes the elevation and azimuth a projectile, a target, a gun and aerodynamic coefficient data.
optional options are the computational time, the wind velocity and the atmosphere.
The trajectory is computed using MPMM

Types Documentation

ExternalBallistics.AbstractTarget — Type

abstract type under which target types are defined.

ExternalBallistics.AbstractPenetrator — Type

abstract type under which projectiles types are defined.

ExternalBallistics.Wind — Type

Wind is defined by the cross and range winds magnitude

ExternalBallistics.Projectile — Type

Projectile is defined by the following parameters: mass, calibre,
positionsion, velocity, tof, Ix, Iy, and Xcg

ExternalBallistics.TargetRect — Type

defines a rectangular target by specifying the lenght a, the height b and the position (range)

ExternalBallistics.TargetCirc — Type

defines a circular target by specifying the radius ρ and the position (range)

ExternalBallistics.Air — Type

defines the atmosphere by specifying the pressure, temperature and relative humidity

ExternalBallistics.Gun — Type

Gun is defined by the muzzle velocity, the latitude, the twist, the gun length, the height, the elevation and azimuth