ExternalBallistics.jl Documentation

Trajectory computation.

Introduction

julia> using ExternalBallistics

Table of contents

ExternalBallistics.jl Documentation

Introduction

Table of contents

Package Features

Functions Documentation

Types Documentation

Package Features

MPMM direct and indirect

Functions Documentation

Objects generation

ExternalBallistics.createTargetRect — Function

createTargetRect(a,b, position) Return a rectangular target by specifying the length 'a', the height 'b' and the 'position' (range)

ExternalBallistics.createTargetCirc — Function

createTargetCirc(size, position)

Returns a circular target by specifying the radius (size) and the position (range)

ExternalBallistics.createTarget — Function

createTarget()

Retursns an object of type 'Target'. Optional arguments are:

ρ the size of the target

position, the range to the target

ExternalBallistics.createGun — Function

createGun(u₀,lat,QE,AZ,tc; args ...)

Returns an object of type 'Gun' by specifying the muzzle velocity (u₀) its latitude (lat), the elevation angle (QE), azimuth angle (AZ) and the twist (tc). Optional arguments are

lw, the gun length

X2w, the height of the gun

Trajectory computation

ExternalBallistics.trajectoryMPMM — Function

trajectoryMPMM(proj, target, gun, aero; args ... )

Returns the impact point (vector), the velocity at impact (vector) and the time of flight of the projectile (proj). The stop condition for the trajectory computation is the range of the target. Launcher conditions are specified by the gun parameters. The projectile aerodynamic coefficient are specified in the aero variable. Optional arguments are:

tspan = the computational time interval. Default is (0.0,1000.0)

R = the earth radius. Default is 6.356766*1e6

Ω = the earth spin rate. Default is 7.292115*1e-5

w_bar = the wind velocity. Default is [0.0,0.0,0.0]

atm =the atmosphere characteristics. Default is nothing

The projectile trajectory is computed using MPMM

ExternalBallistics.QEfinderMPMM! — Function

QEfinderMPMM!(drone, proj, gun,aero;args ...)

Returns the elevation and azimuth departure angles of the projectile (proj) with target (drone). Launcher conditions are specified by the gun parameters, projectile aerodynamic coefficient are specified in the aero variable. Optional arguments are:

tspan = the computational time interval. Default is (0.0,1000.0)

w_bar = the wind velocity. Default is [0.0,0.0,0.0]

atm =the atmosphere characteristics. Default is nothing

The trajectory is computed using MPMM

Types Documentation

ExternalBallistics.AbstractTarget — Type

abstract type under which target types are defined.

ExternalBallistics.AbstractPenetrator — Type

abstract type under which projectiles types are defined.

ExternalBallistics.Wind — Type

Wind(cross, range)

Wind is defined by the:

cross wind magnitude

range wind magnitude

ExternalBallistics.Projectile — Type

Projectile(mass,calibre,position,tof,Ix,Iy,Xcg)

Projectile is defined by the following parameters:

mass (kg)

calibre (m)

positionsion: coordinate of the 'aimpoint'

velocity: velocity vector

tof: time of flight

Ix: axial moment of inertia

Iy: transversal moment of inertia

Xcg: distance to the center of gravity

ExternalBallistics.TargetRect — Type

TargetRect(a,b,position)

defines a rectangular target by specifying the lenght a, the height b and the position (range)

ExternalBallistics.TargetCirc — Type

TargetCirc(ρ,position)

defines a circular target by specifying the radius ρ and the position (range)

ExternalBallistics.Air — Type

Air(p,t,rh)

defines the atmosphere by specifying:

p: the pressure

t: the temperature

rh: the relative humidity

ExternalBallistics.Gun — Type

Gun(u₀,lat,tc,lw,X2w,QE,AZ)

Gun is defined by the following parameters:

u₀: the muzzle velocity

lat: the latitude

tc: the twist

lw: the gun length

X2w: the height

QE: the the elevation angle

AZ: the azimuth