ExternalBallistics.jl Documentation

Trajectory computation.

Introduction

julia> using ExternalBallistics

Table of contents

ExternalBallistics.jl Documentation

Introduction

Table of contents

Package Features

Functions Documentation

Types Documentation

Package Features

MPMM direct and indirect

Functions Documentation

ExternalBallistics.QEfinderMPMM! — Method

QEfinderMPMM!(drone, proj, gun,aero;args ...)

Returns the elevation and azimuth departure angles of the projectile (proj) with target (drone). Launcher conditions are specified by the gun parameters, projectile aerodynamic coefficient are specified in the aero variable. Optional arguments are:

tspan = the computational time interval. Default is (0.0,1000.0)

w_bar = the wind velocity. Default is [0.0,0.0,0.0]

atm =the atmosphere characteristics. Default is nothing

The trajectory is computed using MPMM

ExternalBallistics.createGun — Method

createGun(u₀,lat,QE,AZ,tc; args ...)

Returns an object of type 'Gun' by specifying the muzzle velocity (u₀) its latitude (lat), the elevation angle (QE), azimuth angle (AZ) and the twist (tc). Optional arguments are

lw, the gun length

X2w, the height of the gun

ExternalBallistics.createTarget — Method

createTarget()

Retursns an object of type 'Target'. Optional arguments are:

ρ the size of the target

position, the range to the target

ExternalBallistics.frag — Method

frag(proj, zones, shapes)

Returns the a vector of fragments:

proj : the fragmenting projectile with projectile velocity and position

zones : contains information on the static fragmentation

shapes: contains information on fragments shapes

ExternalBallistics.trajectoryMPMM — Method

trajectoryMPMM(proj, target, gun, aero; args ... )

Returns the impact point (vector), the velocity at impact (vector) and the time of flight of the projectile (proj). The stop condition for the trajectory computation is the range of the target. Launcher conditions are specified by the gun parameters. The projectile aerodynamic coefficient are specified in the aero variable. Optional arguments are:

tspan = the computational time interval. Default is (0.0,1000.0)

R = the earth radius. Default is 6.356766*1e6

Ω = the earth spin rate. Default is 7.292115*1e-5

w_bar = the wind velocity. Default is [0.0,0.0,0.0]

atm =the atmosphere characteristics. Default is nothing

The projectile trajectory is computed using MPMM

Types Documentation

ExternalBallistics.AbstractPenetrator — Type

abstract type under which projectiles types are defined.

ExternalBallistics.AbstractTarget — Type

abstract type under which target types are defined.

ExternalBallistics.Air — Type

Air(p,t,rh)

defines the atmosphere by specifying:

p: the pressure

t: the temperature

rh: the relative humidity

ExternalBallistics.FragShapes — Type

FragShapes(cd, cf)

defines the fragment by specifying:

cd : drag coefficient

cf : shape factor

ExternalBallistics.Fragment — Type

Fragment(position, position, density)

defines the fragment by specifying:

position

position

density

ExternalBallistics.Gun — Type

Gun(u₀,lat,tc,lw,X2w,QE,AZ)

Gun is defined by the following parameters:

u₀: the muzzle velocity

lat: the latitude

tc: the twist

lw: the gun length

X2w: the height

QE: the the elevation angle

AZ: the azimuth

ExternalBallistics.Projectile — Type

Projectile(mass,calibre,position,tof,Ix,Iy,Xcg)

Projectile is defined by the following parameters:

mass (kg)

calibre (m)

positionsion: coordinate of the 'aimpoint'

velocity: velocity vector

tof: time of flight

Ix: axial moment of inertia

Iy: transversal moment of inertia

Xcg: distance to the center of gravity

ExternalBallistics.Wind — Type

Wind(cross, range)

Wind is defined by the:

cross wind magnitude

range wind magnitude

ExternalBallistics.Zone — Type

Zone(angle, μ_number, σ_number, μ_mass, σ_mass, μ_velocity, σ_velocity, density)

defines a framentation zone:

angleBegin: begin angle of the segment (degrees)

angleEnd: end angle of the segment (degrees)

μ_number: average number of fragments

σ_number: standard deviation of the distribution of number of fragments

μ_mass: average mass of the fragments

σ_mass: standard deviation of the distribution of the mass of fragments

μ_velocity: average velocity of the fragments

σ_velocity: standard deviation of the distribution of the velocity of fragments

density: density of the fragments